+++ app/src/main/kotlin/ac/mdiq/podcini/receiver/WidgetReceiver.kt
@@ -1,11 +1,15 @@
 package ac.mdiq.podcini.receiver
 
 import ac.mdiq.podcini.R
+import ac.mdiq.podcini.activity.EpisodeInfoActivity
+import ac.mdiq.podcini.activity.MainActivity
+import ac.mdiq.podcini.activity.PlayerUIActivity
+import ac.mdiq.podcini.activity.QueuePickerActivity
+import ac.mdiq.podcini.activity.starter.MainActivityStarter
 import ac.mdiq.podcini.playback.PlaybackStarter
-import ac.mdiq.podcini.playback.base.InTheatre.actQueue
-import ac.mdiq.podcini.playback.base.InTheatre.ensureAController
 import ac.mdiq.podcini.playback.base.InTheatre.curEpisode
 import ac.mdiq.podcini.playback.base.InTheatre.curState
+import ac.mdiq.podcini.playback.base.InTheatre.ensureAController
 import ac.mdiq.podcini.playback.base.InTheatre.playVideo
 import ac.mdiq.podcini.playback.base.InTheatre.setAsCurEpisode
 import ac.mdiq.podcini.playback.base.MediaPlayerBase.Companion.currentMediaType
@@ -14,36 +18,35 @@ import ac.mdiq.podcini.playback.base.MediaPlayerBase.Companion.isPlaying
 import ac.mdiq.podcini.playback.base.MediaPlayerBase.Companion.mPlayer
 import ac.mdiq.podcini.playback.base.MediaPlayerBase.Companion.playPause
 import ac.mdiq.podcini.playback.base.MediaPlayerBase.Companion.status
-import ac.mdiq.podcini.storage.specs.VideoMode
 import ac.mdiq.podcini.playback.service.PlaybackService.Companion.isCasting
 import ac.mdiq.podcini.playback.service.PlaybackService.Companion.isRunning
 import ac.mdiq.podcini.playback.service.PlaybackService.Companion.playbackService
 import ac.mdiq.podcini.storage.database.episodeById
 import ac.mdiq.podcini.storage.database.fastForwardSecs
+import ac.mdiq.podcini.storage.database.realm
 import ac.mdiq.podcini.storage.database.rewindSecs
 import ac.mdiq.podcini.storage.database.smartRemoveFromAllQueues
 import ac.mdiq.podcini.storage.model.Episode
+import ac.mdiq.podcini.storage.model.PlayQueue
+import ac.mdiq.podcini.storage.model.WidgetEpisode
+import ac.mdiq.podcini.storage.model.toWidget
 import ac.mdiq.podcini.storage.specs.EpisodeState
 import ac.mdiq.podcini.storage.specs.MediaType
 import ac.mdiq.podcini.storage.specs.Rating
+import ac.mdiq.podcini.storage.specs.VideoMode
 import ac.mdiq.podcini.storage.utils.durationStringFull
-import ac.mdiq.podcini.activity.EpisodeInfoActivity
-import ac.mdiq.podcini.activity.MainActivity
-import ac.mdiq.podcini.activity.PlayerUIActivity
-import ac.mdiq.podcini.activity.QueuePickerActivity
-import ac.mdiq.podcini.activity.starter.MainActivityStarter
 import ac.mdiq.podcini.utils.Logd
 import ac.mdiq.podcini.utils.Loge
 import ac.mdiq.podcini.utils.formatDateTimeFlex
-import ac.mdiq.podcini.utils.formatLargeInteger
 import android.content.Context
 import android.content.Intent
-import android.text.format.Formatter
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.datastore.preferences.core.Preferences
+import androidx.datastore.preferences.core.intPreferencesKey
 import androidx.datastore.preferences.core.longPreferencesKey
+import androidx.datastore.preferences.core.stringPreferencesKey
 import androidx.glance.ColorFilter
 import androidx.glance.GlanceId
 import androidx.glance.GlanceModifier
@@ -83,14 +86,16 @@ import androidx.glance.text.Text
 import androidx.glance.text.TextStyle
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
-
+import kotlinx.serialization.json.Json
 
 private const val TAG = "WidgetReceiver"
+
 class WidgetReceiver : GlanceAppWidgetReceiver() {
     override val glanceAppWidget: GlanceAppWidget = PodciniWidget()
 }
 
-private var episodes = listOf<Episode>()
+val episodesJson = stringPreferencesKey("episodes_json")
+
 
 class PodciniWidget : GlanceAppWidget() {
     override val stateDefinition: GlanceStateDefinition<*> = PreferencesGlanceStateDefinition
@@ -100,80 +105,85 @@ class PodciniWidget : GlanceAppWidget() {
     val buttonColorProvider = ColorProvider(day = Color(0xDDFFD700), night = Color(0xDDFFD700))
 
     override suspend fun provideGlance(context: Context, id: GlanceId) {
-        Logd(TAG, "provideGlance id: $id actQueue ${actQueue.name}")
+//        Logd(TAG, "provideGlance id: $id actQueue ${actQueue.name}")
+
+        var episodes: List<WidgetEpisode> = listOf()
 
         provideContent { GlanceTheme {
-            episodes = actQueue.episodesSorted
             Logd(TAG, "provideGlance in provideContent id: $id")
 
             val prefs = currentState<Preferences>()
-            val markedId = prefs[MARKED_EPISODE_KEY]
 
-            Column(modifier = GlanceModifier.fillMaxSize().background(brColorProvider).padding(5.dp), verticalAlignment = Alignment.CenterVertically) {
-                Row(modifier = GlanceModifier.fillMaxWidth()) {
-                    Box(modifier = GlanceModifier.clickable(
-                        actionStartActivity(Intent(context, QueuePickerActivity::class.java).apply { addFlags(Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TOP) }),
-                        rippleOverride = R.drawable.widget_ripple
-                    ).background(ImageProvider(R.drawable.widget_ripple))) {
-                        Text(text = (actQueue.name + ": " + episodes.size), style = TextStyle(color = textColorProvider, fontSize = 18.sp, fontWeight = FontWeight.Bold), maxLines = 1) }
+            // these unfortunately have to be reset every time
+            var markedId = prefs[MARKED_EPISODE_KEY]
+            var queueName = prefs [stringPreferencesKey("queue_name")] ?: "Default"
+            var queueId = prefs[longPreferencesKey("queue_id")] ?: 0L
+            var queueSize = prefs[intPreferencesKey("queue_size")] ?: 0
+
+            val updateYpe = prefs[stringPreferencesKey("update_type")] ?: ""
+            Logd(TAG, "provideGlance updateYpe: $updateYpe")
+
+            when (updateYpe) {
+                "update" -> {
+                    val json = prefs[stringPreferencesKey("episodes")] ?: "[]"
+                    episodes = Json.decodeFromString<List<WidgetEpisode>>(json)
+                }
+                "episode" -> markedId = prefs[MARKED_EPISODE_KEY]
+                "queue" -> {
+                    queueId = prefs[longPreferencesKey("queue_id")] ?: 0L
+                    queueName = prefs [stringPreferencesKey("queue_name")] ?: "Default"
+                    queueSize = prefs[intPreferencesKey("queue_size")] ?: 0
+                    val json = prefs[stringPreferencesKey("episodes")] ?: "[]"
+                    episodes = Json.decodeFromString<List<WidgetEpisode>>(json)
+                    Logd(TAG, "provideGlance episodes: ${episodes.size}")
+                }
+            }
+
+            Column(modifier = GlanceModifier.fillMaxSize().background(brColorProvider).padding(2.dp), verticalAlignment = Alignment.CenterVertically) {
+                Row(modifier = GlanceModifier.fillMaxWidth(), verticalAlignment = Alignment.CenterVertically) {
+                    Box(modifier = GlanceModifier.clickable(actionStartActivity(Intent(context, QueuePickerActivity::class.java).apply { addFlags(Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TOP) }),
+                        rippleOverride = R.drawable.widget_ripple).background(ImageProvider(R.drawable.widget_ripple))) {
+                        Text(text = ("$queueName: $queueSize"), style = TextStyle(color = textColorProvider, fontSize = 18.sp, fontWeight = FontWeight.Bold), maxLines = 1)
+                    }
                     Spacer(GlanceModifier.defaultWeight())
-                    Image(provider = ImageProvider(R.drawable.ic_launcher_foreground), contentDescription = "App",
-                        modifier = GlanceModifier.size(48.dp).clickable(actionStartActivity<MainActivity>(), rippleOverride = R.drawable.widget_ripple))
+                    Image(provider = ImageProvider(R.drawable.ic_launcher_foreground), contentDescription = "App", modifier = GlanceModifier.size(48.dp).clickable(actionStartActivity<MainActivity>(), rippleOverride = R.drawable.widget_ripple).background(ImageProvider(R.drawable.widget_ripple)))
                     Spacer(GlanceModifier.width(10.dp))
-                    Image(provider = ImageProvider(R.drawable.ic_refresh), contentDescription = "Refresh", colorFilter = ColorFilter.tint(textColorProvider),
-                        modifier = GlanceModifier.size(48.dp).clickable(actionRunCallback<RefreshAction>(), rippleOverride = R.drawable.widget_ripple))
+                    Image(provider = ImageProvider(R.drawable.ic_refresh), contentDescription = "Refresh", colorFilter = ColorFilter.tint(textColorProvider), modifier = GlanceModifier.size(48.dp).clickable(actionRunCallback<RefreshAction>(parameters = actionParametersOf(QUEUE_ID_KEY to queueId)), rippleOverride = R.drawable.widget_ripple).background(ImageProvider(R.drawable.widget_ripple)))
                 }
                 LazyColumn(modifier = GlanceModifier.defaultWeight().fillMaxWidth()) {
                     items(episodes) { episode ->
                         Row(modifier = GlanceModifier.fillMaxWidth().padding(bottom = 4.dp), verticalAlignment = Alignment.CenterVertically) {
                             Image(provider = ImageProvider(R.drawable.ic_close_white), contentDescription = "remove", colorFilter = ColorFilter.tint(buttonColorProvider),
-                                modifier = GlanceModifier.size(36.dp).clickable(
-                                    actionRunCallback<RemoveAction>(parameters = actionParametersOf(EPISODE_ID_KEY to episode.id)),
-                                    rippleOverride = R.drawable.widget_ripple
-                                ))
+                                modifier = GlanceModifier.size(36.dp).clickable(actionRunCallback<RemoveAction>(parameters = actionParametersOf(EPISODE_ID_KEY to episode.id, QUEUE_ID_KEY to queueId)), rippleOverride = R.drawable.widget_ripple).background(ImageProvider(R.drawable.widget_ripple)))
                             val isMarked = episode.id == markedId || episode.id == curEpisode?.id
-                            Column(modifier = GlanceModifier.defaultWeight().clickable(
-                                actionStartActivity<EpisodeInfoActivity>(parameters = actionParametersOf(EPISODE_INFO_ID_KEY to episode.id)),
-                                rippleOverride = R.drawable.widget_ripple
-                            )) {
-                                Text(episode.title ?: "", style = TextStyle(color = textColorProvider, fontSize = 13.sp, fontWeight = if (isMarked) FontWeight.Bold else FontWeight.Normal), maxLines = 2)
+                            Column(modifier = GlanceModifier.defaultWeight().clickable(actionStartActivity<EpisodeInfoActivity>(parameters = actionParametersOf(EPISODE_INFO_ID_KEY to episode.id)), rippleOverride = R.drawable.widget_ripple).background(ImageProvider(R.drawable.widget_ripple))) {
+                                Text(episode.t ?: "", style = TextStyle(color = textColorProvider, fontSize = 13.sp, fontWeight = if (isMarked) FontWeight.Bold else FontWeight.Normal), maxLines = 1)
                                 Row(verticalAlignment = Alignment.CenterVertically) {
-                                    Image(provider = ImageProvider(EpisodeState.fromCode(episode.playState).res), contentDescription = "playState", colorFilter = ColorFilter.tint(buttonColorProvider), modifier = GlanceModifier.size(16.dp))
-                                    if (episode.rating != Rating.UNRATED.code) Image(provider = ImageProvider(Rating.fromCode(episode.rating).res), contentDescription = "rating", colorFilter = ColorFilter.tint(buttonColorProvider), modifier = GlanceModifier.size(16.dp))
-                                    val dateSizeText = " · " + formatDateTimeFlex(episode.pubDate) + " · " + durationStringFull(episode.duration) + (if (episode.size > 0) " · " + Formatter.formatShortFileSize(context, episode.size) else "") + (if (episode.viewCount > 0) " · " + formatLargeInteger(episode.viewCount) else "")
+                                    Image(provider = ImageProvider(EpisodeState.fromCode(episode.s).res), contentDescription = "playState", colorFilter = ColorFilter.tint(buttonColorProvider), modifier = GlanceModifier.size(16.dp))
+                                    if (episode.r != Rating.UNRATED.code) Image(provider = ImageProvider(Rating.fromCode(episode.r).res), contentDescription = "rating", colorFilter = ColorFilter.tint(buttonColorProvider), modifier = GlanceModifier.size(16.dp))
+                                    val dateSizeText = " · " + formatDateTimeFlex(episode.pd) + " · " + durationStringFull(episode.du)
                                     Text(dateSizeText, style = TextStyle(color = textColorProvider, fontSize = 10.sp), maxLines = 1)
                                 }
                             }
                             Image(provider = ImageProvider(R.drawable.outline_play_pause_24), contentDescription = "Play/pause", colorFilter = ColorFilter.tint(if (isMarked) textColorProvider else buttonColorProvider),
-                                modifier = GlanceModifier.size(48.dp).clickable(
-                                    actionRunCallback<PlayAction>(parameters = actionParametersOf(EPISODE_ID_KEY to episode.id)),
-                                    rippleOverride = R.drawable.widget_ripple
-                                ))
+                                modifier = GlanceModifier.size(48.dp).clickable(actionRunCallback<PlayAction>(parameters = actionParametersOf(EPISODE_ID_KEY to episode.id)), rippleOverride = R.drawable.widget_ripple).background(ImageProvider(R.drawable.widget_ripple)))
                         }
                     }
                 }
                 val buttonSize = 60.dp
-                Row(modifier = GlanceModifier.fillMaxWidth().padding(horizontal = 10.dp),  horizontalAlignment = Alignment.CenterHorizontally) {
+                Row(modifier = GlanceModifier.fillMaxWidth().padding(horizontal = 5.dp),  horizontalAlignment = Alignment.CenterHorizontally) {
                     Image(provider = ImageProvider(R.drawable.outline_smart_display_24), contentDescription = "PlayerUI", colorFilter = ColorFilter.tint(buttonColorProvider),
-                        modifier = GlanceModifier.size(buttonSize).padding(end = 5.dp).clickable(
-                            actionStartActivity<PlayerUIActivity>(),
-//                            actionStartActivity(Intent(context, PlayerUIActivity::class.java).apply { addFlags(Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TOP) }),
-//                            actionStartActivity(Intent(context, PlayerUIActivity::class.java).apply {
-//                                addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
-//                                addFlags(Intent.FLAG_ACTIVITY_SINGLE_TOP)
-//                                addFlags(Intent.FLAG_ACTIVITY_REORDER_TO_FRONT)                            }),
-                            rippleOverride = R.drawable.widget_ripple
-                        ))
+                        modifier = GlanceModifier.size(buttonSize).padding(end = 5.dp).clickable(actionStartActivity<PlayerUIActivity>(), rippleOverride = R.drawable.widget_ripple).background(ImageProvider(R.drawable.widget_ripple)))
                     Image(provider = ImageProvider(R.drawable.baseline_skip_previous_24), contentDescription = "Restart", colorFilter = ColorFilter.tint(buttonColorProvider),
-                        modifier = GlanceModifier.size(buttonSize).padding(end = 5.dp).clickable(actionRunCallback<PrevAction>(), rippleOverride = R.drawable.widget_ripple))
+                        modifier = GlanceModifier.size(buttonSize).padding(end = 5.dp).clickable(actionRunCallback<PrevAction>(), rippleOverride = R.drawable.widget_ripple).background(ImageProvider(R.drawable.widget_ripple)))
                     Image(provider = ImageProvider(R.drawable.ic_fast_rewind), contentDescription = "Rewind", colorFilter = ColorFilter.tint(buttonColorProvider),
-                        modifier = GlanceModifier.size(buttonSize).padding(end = 5.dp).clickable(actionRunCallback<RewindAction>(), rippleOverride = R.drawable.widget_ripple))
+                        modifier = GlanceModifier.size(buttonSize).padding(end = 5.dp).clickable(actionRunCallback<RewindAction>(), rippleOverride = R.drawable.widget_ripple).background(ImageProvider(R.drawable.widget_ripple)))
                     Image(provider = ImageProvider(R.drawable.outline_play_pause_24), contentDescription = "Play/pause", colorFilter = ColorFilter.tint(buttonColorProvider),
-                        modifier = GlanceModifier.size(buttonSize).padding(end = 5.dp).clickable(actionRunCallback<ToggleAction>(), rippleOverride = R.drawable.widget_ripple))
+                        modifier = GlanceModifier.size(buttonSize).padding(end = 5.dp).clickable(actionRunCallback<ToggleAction>(if (episodes.isNotEmpty()) actionParametersOf(EPISODE_ID_KEY to episodes[0].id) else actionParametersOf()), rippleOverride = R.drawable.widget_ripple).background(ImageProvider(R.drawable.widget_ripple)))
                     Image(provider = ImageProvider(R.drawable.ic_fast_forward), contentDescription = "Forward", colorFilter = ColorFilter.tint(buttonColorProvider),
-                        modifier = GlanceModifier.size(buttonSize).padding(end = 5.dp).clickable(actionRunCallback<ForwardAction>(), rippleOverride = R.drawable.widget_ripple))
+                        modifier = GlanceModifier.size(buttonSize).padding(end = 5.dp).clickable(actionRunCallback<ForwardAction>(), rippleOverride = R.drawable.widget_ripple).background(ImageProvider(R.drawable.widget_ripple)))
                     Image(provider = ImageProvider(R.drawable.ic_skip_48dp), contentDescription = "Skip", colorFilter = ColorFilter.tint(buttonColorProvider),
-                        modifier = GlanceModifier.size(buttonSize).clickable(actionRunCallback<NextAction>(), rippleOverride = R.drawable.widget_ripple))
+                        modifier = GlanceModifier.size(buttonSize).clickable(actionRunCallback<NextAction>(), rippleOverride = R.drawable.widget_ripple).background(ImageProvider(R.drawable.widget_ripple)))
                 }
             }
         } }
@@ -182,26 +192,28 @@ class PodciniWidget : GlanceAppWidget() {
 
 val EPISODE_INFO_ID_KEY = ActionParameters.Key<Long>("episode_info_id")
 val EPISODE_ID_KEY = ActionParameters.Key<Long>("episode_id")
+val QUEUE_ID_KEY = ActionParameters.Key<Long>("queue_id")
+
 val MARKED_EPISODE_KEY = longPreferencesKey("marked_episode_id")
 
-class PlayAction : ActionCallback {
+class RefreshAction : ActionCallback {
     override suspend fun onAction(context: Context, glanceId: GlanceId, parameters: ActionParameters) {
-        Logd(TAG, "onReceive")
-        ensureAController()
-        updateAppWidgetState(context, glanceId) { prefs ->
-            val id = parameters[EPISODE_ID_KEY]
-            if (id == null) {
-                Loge("PlayAction", "id from parameter is null.")
-                return@updateAppWidgetState
+        Logd(TAG, "RefreshAction onAction")
+        val queueId = parameters[QUEUE_ID_KEY]
+        if (queueId == null) {
+            Loge("RefreshAction", "queueId from parameter is null.")
+            return
+        }
+        val episodes = withContext(Dispatchers.IO) {
+            val queue = realm.query(PlayQueue::class).query("id == $queueId").first().find()
+            queue?.episodesSorted?.take(40)?.map { it.toWidget() } ?: listOf()
+        }
+        val json = Json.encodeToString(episodes)
+        updateAppWidgetState(context, PreferencesGlanceStateDefinition, glanceId) { prefs ->
+            prefs.toMutablePreferences().apply {
+                this[stringPreferencesKey("episodes")] = json
+                this[stringPreferencesKey("update_type")] = "update"
             }
-            val episode = episodeById(id)
-            if (episode == null) {
-                Loge("PlayAction", "episode with id: $id is null.")
-                return@updateAppWidgetState
-            }
-            prefs[MARKED_EPISODE_KEY] = id
-            Logd(TAG, "PlayAction onAction episode: ${episode.title}")
-            withContext(Dispatchers.Main) { PlaybackStarter(episode).setWidgetId(glanceId.toString()).shouldStreamThisTime(null).start() }
         }
         PodciniWidget().update(context, glanceId)
     }
@@ -220,18 +232,64 @@ class RemoveAction : ActionCallback {
             return
         }
         Logd(TAG, "RemoveAction onAction episode: ${episode.title}")
-        smartRemoveFromAllQueues(episode)
+        val episodes = withContext(Dispatchers.IO) {
+            smartRemoveFromAllQueues(episode)
+            val queueId = parameters[QUEUE_ID_KEY]
+            if (queueId == null) {
+                Loge("RemoveAction", "queueId from parameter is null.")
+                return@withContext listOf()
+            }
+            Logd(TAG, "RemoveAction onAction queueId: $queueId")
+            val queue = realm.query(PlayQueue::class).query("id == $queueId").first().find()
+            queue?.episodesSorted?.take(40)?.map { it.toWidget() } ?: listOf()
+        }
+        val json = Json.encodeToString(episodes)
+        updateAppWidgetState(context, PreferencesGlanceStateDefinition, glanceId) { prefs ->
+            prefs.toMutablePreferences().apply {
+                this[stringPreferencesKey("episodes")] = json
+                this[stringPreferencesKey("update_type")] = "update"
+            }
+        }
         PodciniWidget().update(context, glanceId)
     }
 }
 
+class PlayAction : ActionCallback {
+    override suspend fun onAction(context: Context, glanceId: GlanceId, parameters: ActionParameters) {
+        Logd(TAG, "onReceive")
+        ensureAController()
+        updateAppWidgetState(context, glanceId) { prefs ->
+            val id = parameters[EPISODE_ID_KEY]
+            if (id == null) {
+                Loge("PlayAction", "id from parameter is null.")
+                return@updateAppWidgetState
+            }
+            val episode = episodeById(id)
+            if (episode == null) {
+                Loge("PlayAction", "episode with id: $id is null.")
+                return@updateAppWidgetState
+            }
+            prefs[MARKED_EPISODE_KEY] = id
+            prefs[stringPreferencesKey("update_type")] = "episode"
+            Logd(TAG, "PlayAction onAction episode: ${episode.title}")
+            withContext(Dispatchers.Main) { PlaybackStarter(episode).setWidgetId(glanceId.toString()).shouldStreamThisTime(null).start() }
+        }
+        PodciniWidget().update(context, glanceId)
+    }
+}
 
 class ToggleAction : ActionCallback {
     override suspend fun onAction(context: Context, glanceId: GlanceId, parameters: ActionParameters) {
         Logd(TAG, "onReceive")
         ensureAController()
         Logd(TAG, "ToggleAction onAction isPlaying: $isPlaying")
-        if (curEpisode == null && episodes.isNotEmpty()) setAsCurEpisode(episodes[0])
+        if (curEpisode == null) {
+            val id = parameters[EPISODE_ID_KEY]
+            if (id != null) {
+                val e = realm.query(Episode::class).query("id == $id").first().find()
+                if (e != null) setAsCurEpisode(e)
+            }
+        }
         if (curEpisode != null) {
             fun getPlayerActivityIntent(context: Context, mediaType_: MediaType? = null): Intent {
                 val mediaType = mediaType_ ?: currentMediaType
@@ -248,16 +306,8 @@ class ToggleAction : ActionCallback {
                     PlaybackStarter(curEpisode!!).setWidgetId(glanceId.toString()).shouldStreamThisTime(null).start()
                 }
             }
+            PodciniWidget().update(context, glanceId)
         }
-        PodciniWidget().update(context, glanceId)
-    }
-}
-
-class RefreshAction : ActionCallback {
-    override suspend fun onAction(context: Context, glanceId: GlanceId, parameters: ActionParameters) {
-        Logd(TAG, "RefreshAction onAction")
-        episodes = actQueue.episodesSorted
-        PodciniWidget().update(context, glanceId)
     }
 }