+++ app/src/main/kotlin/ac/mdiq/podcini/storage/model/PlayQueue.kt
@@ -122,7 +122,7 @@ class PlayQueue : RealmObject {
 
     fun checkAndFill() {
         Logd(TAG, "checkAndFill")
-        showStackTrace()
+//        showStackTrace()
         if (size() == 0 && !isVirtual()) {
             CoroutineScope(Dispatchers.IO).launch {
                 val feeds = normalFeeds